# backend/src/swagger.json
{
  "info": {
    "title": "Construction Bidding API",
    "version": "1.0.0",
    "description": "Test Express API with autogenerated swagger doc"
  },
  "swagger": "2.0",
  "paths": {
    "/project": {
      "get": {
        "description": "Get all projects",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "response": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a new project",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "project",
            "description": "Project object",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "response": {
              "schema": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        }
      }
    },
    "/project/{id}": {
      "patch": {
        "description": "Update a project",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "project",
            "description": "To be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Project"
            }
          },
          {
            "in": "path",
            "name": "id",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "response": {
              "schema": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete a project",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "response": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "description": "Get all current users",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "response": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "description": "Authenticate a magic link for a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "description": "User email address",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "response": {
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "patch": {
        "description": "Update a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "To be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          {
            "in": "path",
            "name": "id",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "response": {
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "response": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Project": {
      "type": "object",
      "properties": {
        "owner_id": {
          "type": "string",
          "example": "5e4cccc8007d3554fd089599"
        },
        "name": {
          "type": "string",
          "example": "Golden Gate Bridge Renovation"
        },
        "status": {
          "type": "string",
          "example": "open",
          "enum": [
            "open",
            "closed"
          ]
        },
        "duration": {
          "type": "number",
          "example": 100
        },
        "budget": {
          "type": "number",
          "example": 1000000
        },
        "description": {
          "type": "string",
          "example": "Bridge must be retrofitted to withstand an 8.0 earthquake"
        },
        "bids": {
          "type": "array",
          "contractor_id": "string",
          "duration": "number",
          "budget": "number"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "me@domain.com"
        },
        "company_name": {
          "type": "string",
          "example": "John"
        },
        "role": {
          "type": "string",
          "example": "owner",
          "enum": [
            "owner",
            "contractor"
          ]
        }
      }
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}